pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.16",
    "fastapi[standard]>=0.115.12",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-mock>=3.14.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "python-telegram-bot==21.0.1",
    "uvicorn[standard]>=0.34.0",
]


# pyproject.toml

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "function"
